themes: Add `ruled.notification` to all themes.
The rational behind this is that adding more and more `beautiful`
variables doesn't scale. While not everybody is on board with this,
I think we should move more logic into the theme to put an end to
the unscalable complexity plagging some modules (such as the
tasklist).

Instead of adding a variable for the background and foreground color
of urgent notification, we can now use the rules.