Move client resize on show titlebar to its method.
For fome reason, the titlebar constructor always works with the
position sets to 'top'. Even if the targeted titlebar is not the top
one.
This is either, a bug that is then corrected by the titlebar geometries
and internal position property ; or the actual normal behavior and I
still can't understand what happen...

Anyway, best way to correctly resize the client is to resize it in the
`:show` method, so we can have both, the titlebar size and the correct
position.