feat(w.w.slider): Add signals for the dragging state
When trying to hook up the slider to an external value, one would
usually have two data streams:

- `property::value` -> set external value
- external value changed -> `.value = new_value`

The problem is that without manual intervention, these two streams form
a loop, as setting `.value` also emits `property::value`.

The new set of signals is disconnected from the `value` property and its
signal and allows for more fine grained inspection of the dragging
state.

Signed-off-by: Lucas Schwiderski <lucas@lschwiderski.de>