container: Add a broder container.
Rather than bloat the `background` container with another 1k lines,
a new container is added. The goal is to implement CSS-style slicing
for the background images. The container also allows to place
generic widgets on each sides and corners of the container.

The main use case for this container is to replace the old
`awful.titlebar` API with `awful.decoration`. That new module will
allow a central client widget to be surrounded by widgets rather than
have 4 separate drawing areas. The border container is designed to
make complex border+titlebar setup trivial.