fix(popup):  rename set_hide_on_right_click
The implementation of the `hide_fct`actually  doesn't check for the
button pressed. And it sounds more user-friendly that it works this way.
So, it is reasonable to rename the constructor parameter to
`hide_on_click`, and also apply this same naming update to the setter.

Note that the changes here should  preserve full backward compatibility
with deprecation warning.