test: Enforce the maximum size for the `wibox` test example templte.
So far, it worked fine if both `w`/`h` were enforced or neither.
If one was, then there was a chance one axis would end up 9999
pixel wide or tall. This worked fine when the `.svg` was inserted
using `<img>`, but now that it's using `<object>`, it now scales
down to zero.