fix(ldoc): set `format` to "markdown" (#3500)
The CI currently doesn't install (correctly?) the "discount" formatter,
so ldoc fallbacks to the default "markdown" formatter. However, if we
try to install the discount or lua-discount, the doc rendering is
slightly different from what we currently have.

All of this was already mentioned in the issue "Generate apidoc doesn't
use the discount markdown formatter #3500".

This change removes the configuration in the `config.ld` file, so we
don't have the "format: using built-in markdown" message any longer in
in the CI logs.

Note that I also removed the lua-discount installation from the
`apidoc.yml` workflow, since it wasn't working. (I guess it installs the lua
package for a different version of lua than the one used by ldoc)