Move the C coverage to the CodeCov GitHub Action
Same as Last PR. Now I expect the Lua coverage to work and the C coverage. I did it in 2 PRs to make sure the delta works.

I also added a `-j8` in the example test step because it is I/O bound and it has enough ram. This should speed-up the CI (maybe?).