awful.menu: map KP_Enter similar to Return (#3979)
In Awesome's menu system, the KP_Enter key is currently used the same
as the Right key (to descend into a child menu). This is unexpected,
since in general, users expect KP_Enter to work similar to the Return
key (to run the highlighted action/command). This commit corrects that
behavior.

Related: https://github.com/awesomeWM/awesome/issues/3867

Signed-off-by: Sicelo A. Mhlongo <absicsz@gmail.com>